+++ SRP/Code/ControlRoom/StandardModuleRibbons.cs
@@ -148,6 +148,7 @@ namespace GRA.SRP.Core.Utilities
             };
             pnl.Add(new RibbonLink { Name = "Event Codes", Url = "/ControlRoom/Modules/Reports/EventCodes.aspx" });
             pnl.Add(new RibbonLink { Name = "Event Issues", Url = "/ControlRoom/Modules/Reports/EventIssues.aspx" });
+            pnl.Add(new RibbonLink { Name = "Challenge Issues", Url = "/ControlRoom/Modules/Reports/ChallengeIssues.aspx" });
 
             //pnl.Add(new RibbonLink { Name = "Game Level Stats", Url = "/ControlRoom/Modules/Reports/LevelStats.aspx" });
             //pnl.Add(new RibbonLink { Name = "Prizes Stats", Url = "/ControlRoom/Modules/Reports/PrizesStats.aspx" });