+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 ### Added
+- Challenge Issues report with Excel export
 - QR Code generation in the Control Room for Events with secret codes
 - Event Issues report with Excel export
 - Event Codes report with Excel export